Move location code functions to their own class
