+++ core_data_modules/cleaners/location_tools/__init__.py
@@ -1,0 +1,1 @@
+from .somalia_locations import SomaliaLocations
